// apps/web/src/api/generated/model/chatDetailDtoTitleSource.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Open WebUI Klon API
 * OpenAPI spec version: 0.0.0
 */

export type ChatDetailDtoTitleSource = typeof ChatDetailDtoTitleSource[keyof typeof ChatDetailDtoTitleSource];


export const ChatDetailDtoTitleSource = {
  fallback: 'fallback',
  generated: 'generated',
  user: 'user',
} as const;